requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
requests==2.31.0
selenium==3.141.0
urllib3==1.25.6


allure-python-commons~=2.13.5